fixes ➡️ in between⬅️

+++ ft_inbetween.c
@@ -6,13 +6,13 @@
 /*   By: apinto <apinto@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/27 08:18:50 by apinto            #+#    #+#             */
-/*   Updated: 2021/04/27 08:19:31 by apinto           ###   ########.fr       */
+/*   Updated: 2021/05/08 13:45:29 by apinto           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_inbetween(int value, int min, int max)
+int	ft_inbetween(double value, double min, double max)
 {
 	return (value >= min && value <= max);
 }
